refactor!: State management overhaul
I've intentionally chosen to move away from the provider package here -
I know that it's the recommended way to manage state, but this app feels
simple enough that managing a few callbacks and staying as close to
"vanilla" flutter as possible is worthwhile. I may walk this choice back
in the future and move back to one of the many abstraction libraries.